test_asserts: Allow % in macro parameters
Previously, when using the test asserts with a %, e.g.
"TEST_EQ(0, 8 % 4)", the build would fail because the "% " would be
interpreted as an invalid format string for snprintf. This commit causes
test macro parameters to provided stringified as parameters to snprintf
instead of leaving them placed stringified in the format string.