Allow build-time toggling of time measurement
Introduce the TIME_MEASUREMENT define that controls whether or not time
measuring and reporting will be done. Define it to 0 or leave it
undefined to avoid time measuring and reporting. Define it to non-zero
to enable time measuring and reporting.

This was added to make the dependency on time.h optional. This allows
this test framework to run even on platforms that don't have a way of
keeping track of time. For those platforms that do have a way to measure
time, and a time.h, defining TIME_MEASUREMENT to non-zero will retain
the old, time measuring and reporting behavior.