junit_xml: Allow disabling of JUnit XML
Sometimes we don't want to depend on file reading and writing portions
of the C library. By allowing the compile-time disabling of JUnit XML
output, we can avoid having a dependency on those parts of the C
library. This is useful when running in an environment without a C
library.