Remove dependency on setjmp and longjmp
Use a global variable 'test_passed' (instead of a global jmp_buf) to
keep track of whether or not a test run passes or fails. In doing this,
we can avoid the use of setjmp and longjmp, which are not present in
environments with a limited or non-existent C library.